Two projects
Two short programs written. Not bad, got all of one without needing help, the second one was a bit more difficult in getting the operations correct, I was using <= when I just needed <, and just needed the last else statement as a catch all for the other states not covered in the first if, else if statements.